apply some changes
