+++ tesseract_core/sdk/templates/pytorch/tesseract_api.py
@@ -4,18 +4,15 @@
 # Tesseract API module for {{name}}
 # Generated by tesseract {{version}} on {{timestamp}}
 
-from typing import Any, Callable
+from typing import Any
 
 import numpy as np
 import torch
 from pydantic import BaseModel
-from torch.utils._pytree import TreeSpec, tree_flatten, tree_unflatten
+from torch.utils._pytree import tree_map
 
 from tesseract_core.runtime import Differentiable, Float32
-from tesseract_core.runtime.tree_transforms import (
-    flatten_with_paths,
-    set_at_path,
-)
+from tesseract_core.runtime.tree_transforms import filter_pos_func, flatten_with_paths
 
 #
 # Schemata
@@ -47,9 +44,6 @@ def evaluate(inputs: Any) -> Any:
 
 
 def apply(inputs: InputSchema) -> OutputSchema:
-    # Convert to pytorch tensors to enable torch.jit
-    inputs = convert_to_tensors(inputs.model_dump())
-
     # Optional: Insert any pre-processing/setup that doesn't require tracing
     # and is only required when specifically running your apply function
     # and not your differentiable endpoints.
@@ -57,6 +51,9 @@ def apply(inputs: InputSchema) -> OutputSchema:
     # Pre-processing should not modify any input that could impact the
     # differentiable outputs in a nonlinear way (a constant shift
     # should be safe)
+
+    # Convert to pytorch tensors
+    inputs = tree_map(convert_to_tensors, inputs.model_dump())
 
     out = evaluate(inputs)
 
@@ -77,17 +74,18 @@ def jacobian(
     jac_outputs: set[str],
 ):
     # convert all numbers and arrays to torch tensors
-    tensor_inputs = convert_to_tensors(inputs.model_dump())
+    tensor_inputs = tree_map(convert_to_tensors, inputs.model_dump())
 
     # flatten the dictionaries such that they can be accessed by paths
     path_inputs = flatten_with_paths(tensor_inputs, jac_inputs)
 
     # transform the dictionaries into a list of values for a positional function
-    pos_inputs, treedef = tree_flatten(path_inputs)
+    pos_inputs = path_inputs.values()
+    keys = path_inputs.keys()
 
     # create a positional function that accepts a list of values and returns a set of tuples
     filtered_pos_eval = filter_pos_func(
-        evaluate, tensor_inputs, jac_outputs, treedef, True
+        evaluate, tensor_inputs, jac_outputs, keys, True
     )
 
     # calculate the jacobian
@@ -110,20 +108,22 @@ def jacobian_vector_product(
     tangent: dict[str, Any],
 ):
     # convert all numbers and arrays to torch tensors
-    tensor_inputs = convert_to_tensors(inputs.model_dump())
-    tensor_tangent = convert_to_tensors(tangent)
+    tensor_inputs = tree_map(convert_to_tensors, inputs.model_dump())
+    tensor_tangent = tree_map(convert_to_tensors, tangent)
 
     # flatten the dictionaries such that they can be accessed by paths
     path_inputs = flatten_with_paths(tensor_inputs, jvp_inputs)
 
     # transform the dictionaries into a list of values for a positional function
-    pos_inputs, treedef = tree_flatten(path_inputs)
-    pos_tangent, _ = tree_flatten(tensor_tangent)
+    pos_inputs = path_inputs.values()
+    keys_inputs = path_inputs.keys()
 
-    # sort
+    pos_tangent = tensor_tangent.values()
 
     # create a positional function that accepts a list of values
-    filtered_pos_eval = filter_pos_func(evaluate, tensor_inputs, jvp_outputs, treedef)
+    filtered_pos_eval = filter_pos_func(
+        evaluate, tensor_inputs, jvp_outputs, keys_inputs
+    )
 
     tangent = torch.func.jvp(filtered_pos_eval, tuple(pos_inputs), tuple(pos_tangent))
 
@@ -138,26 +138,23 @@ def vector_jacobian_product(
 ):
     # Make ordering of vjp in and output args deterministic
     # Necessacy as torch.vjp function requires inputs and outputs to be in the same order
-    vjp_inputs = list(vjp_inputs)
-    vjp_outputs = list(vjp_outputs)
-    vjp_inputs.sort()
-    vjp_outputs.sort()
-
-    # sort the cotangent vector
     cotangent_vector = {key: cotangent_vector[key] for key in vjp_outputs}
 
     # convert all numbers and arrays to torch tensors
-    tensor_inputs = convert_to_tensors(inputs.model_dump())
-    tensor_cotangent = convert_to_tensors(cotangent_vector)
+    tensor_inputs = tree_map(convert_to_tensors, inputs.model_dump())
+    tensor_cotangent = tree_map(convert_to_tensors, cotangent_vector)
 
     # flatten the dictionaries such that they can be accessed by paths
     path_inputs = flatten_with_paths(tensor_inputs, vjp_inputs)
 
     # transform the dictionaries into a list of values for a positional function
-    pos_inputs, treedef = tree_flatten(path_inputs)
+    pos_inputs = path_inputs.values()
+    keys_inputs = path_inputs.keys()
 
     # create a positional function that accepts a list of values
-    filtered_pos_func = filter_pos_func(evaluate, tensor_inputs, vjp_outputs, treedef)
+    filtered_pos_func = filter_pos_func(
+        evaluate, tensor_inputs, vjp_outputs, keys_inputs
+    )
 
     _, vjp_func = torch.func.vjp(filtered_pos_func, *pos_inputs)
 
@@ -171,39 +168,6 @@ def vector_jacobian_product(
     return res_dict
 
 
-def filter_pos_func(
-    func: Callable[[dict], dict],
-    default_inputs: dict,
-    output_paths: set[str],
-    pytree: TreeSpec,
-    output_to_tuple: bool = False,
-) -> Callable:
-    """Returns a reduced func with default inputs that operates on positional arguments.
-
-    The returned function will accept a tuple of positional arguments,
-    convert them back to a dictionary and update the default inputs
-    with the new values at each path. It will then call the original function with the updated inputs
-    and return a dictionary `{output_path: value}`.
-    """
-
-    # function that accepts positional arguments
-    def filtered_pos_func(*args):
-        # convert back to dictionary
-        new_inputs = tree_unflatten(args, pytree)
-
-        # partially update the default inputs with the new values
-        updated_inputs = set_at_path(default_inputs, new_inputs)
-
-        path_outputs = flatten_with_paths(func(updated_inputs), output_paths)
-
-        if output_to_tuple:
-            return tuple(tree_flatten(path_outputs)[0])
-
-        return path_outputs
-
-    return filtered_pos_func
-
-
 def convert_to_tensors(data):
     """Convert all numbers and arrays to torch tensors."""
     if isinstance(data, np.ndarray):
@@ -214,21 +178,5 @@ def convert_to_tensors(data):
         return torch.tensor(data)
     elif isinstance(data, (np.bool_, bool)):
         return torch.tensor(data)
-    elif isinstance(data, dict):
-        return {key: convert_to_tensors(value) for key, value in data.items()}
-    elif isinstance(data, list):
-        return [convert_to_tensors(item) for item in data]
     else:
         raise TypeError(f"Unsupported data type: {type(data)}")
-
-
-def convert_to_numpy(data):
-    """Convert all numbers and arrays to numpy arrays."""
-    if isinstance(data, torch.Tensor):
-        return data.detach().numpy()
-    elif isinstance(data, dict):
-        return {key: convert_to_numpy(value) for key, value in data.items()}
-    elif isinstance(data, list):
-        return [convert_to_numpy(item) for item in data]
-    else:
-        return data
